+++ src/lib.rs
@@ -39,6 +39,7 @@ pub mod sidx;
 pub mod styp;
 pub mod timecode;
 pub mod track;
+pub mod track_group;
 pub mod track_load;
 pub mod track_selection;
 pub mod user_data;
@@ -126,6 +127,9 @@ pub use timecode::{
     TMCD_FLAG_DROP_FRAME, TMCD_FLAG_NEGATIVES_OK,
 };
 pub use track::{SampleDescription, Track, TrackRef, TrackRefKind};
+pub use track_group::{
+    parse_track_group_type, parse_trgr, TrackGroupTypeEntry, TRACK_GROUP_TYPE_MSRC,
+};
 pub use track_load::{
     parse_load, Load, LOAD_HINT_DOUBLE_BUFFER, LOAD_HINT_HIGH_QUALITY, LOAD_PRELOAD_ALWAYS,
     LOAD_PRELOAD_DURATION_TO_END, LOAD_PRELOAD_IF_ENABLED,